test/support/builders: Rename ./pouchdb to ./db
Since those can be used for both CouchDB and PouchDB data.
